Now applies the root bone axis flags from the dominating action. (Eventually we'll have these stores per animation instead of action, still on the todo list.)
Improved UnitScale calculation so root motion is really slick now.

+++ src/Shared/EntitySkeleton.h
@@ -253,7 +253,7 @@ struct EntitySkeletonBlendActionState {
 	//! 'Old' frame.
 	int32_t oldFrame = 0;
 	//! BackLerp.
-	float backLerp = 0.f;
+	double backLerp = 0.f;
 };
 
 struct EntitySkeleton {
